BulkProcessor: First rough draft
This is a first draft of a BulkProcessor similar to the one available
in the Java API [1], [2].

The purpose of a BulkProcessor is to accept BulkableRequests (Index,
Update, or Delete) and commit them based on a policy. The policy is
either if the number of requests exceed a certain limit, exceed a
certain (estimated) byte size, or were not committed during a given
interval.

This should be excessively tested. It will also probably not be the
final API.

[1]
https://github.com/elastic/elasticsearch/blob/2.x/core/src/main/java/org/elasticsearch/action/bulk/BulkProcessor.java
[2]
https://www.elastic.co/guide/en/elasticsearch/client/java-api/current/java-docs-bulk-processor.html