Nodes need to be awaken if all are dead
If we don't rely on sniffing to find healthy nodes, a client might be
stuck with all nodes dead and never return.

So as a resort, we mark all nodes as alive in that case. The next call
to `PerformRequest` should then find a node that is alive (not healthy)
and try again. If any node came back again, it will be marked as
healthy again, eventually.

See #192.