Adding retries and refactoring
This commit adds retries to the workers. They will now try several
times to send bulk requests to Elasticsearch before giving up.

Also, the Failure handler was removed: Only Before and After callbacks
are left. The Failure and the After handlers were merged: After now
comes with an error parameter that indicates whether the given batch
was successful or not.

Also documented the current state of bulk processor in the Wiki:
https://github.com/olivere/elastic/wiki/BulkProcessor